+++ CHANGELOG.md
@@ -7,6 +7,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ### Added
 - ZonedDateTimes methods to convert between ZDT and Sql Timestamps
+- EnveloperFactory method to create a Enveloper with a list of Event classes registered
+- EventStreamMockHelper that provides a method to verify and return the Stream argument from a call 
+to the EventStream append method
 
 ### Changed
 - Event Log to include a 'date_created' timestamp field (Event Store).